// src/views/tokens.d.ts
declare interface DesignToken {
  value: any;
  name?: string;
  comment?: string;
  themeable?: boolean;
  attributes?: {
    category?: string;
    type?: string;
    item?: string;
    subitem?: string;
    state?: string;
    [key: string]: any;
  };
  [key: string]: any;
}

declare const tokens: {
  "color": {
    "palette": {
      "white": DesignToken,
      "grey": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "red": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "orange": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "yellow": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "lime": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "green": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "teal": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "turquoise": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "aqua": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "blue": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "ultramarine": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "purple": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      },
      "pink": {
        "100": DesignToken,
        "090": DesignToken,
        "080": DesignToken,
        "070": DesignToken,
        "060": DesignToken,
        "050": DesignToken,
        "040": DesignToken,
        "030": DesignToken,
        "020": DesignToken,
        "010": DesignToken
      }
    },
    "mui": {
      "palette": {
        "primary": DesignToken,
        "primaryDark": DesignToken,
        "secondary": DesignToken,
        "secondaryDark": DesignToken,
        "error": DesignToken,
        "errorDark": DesignToken,
        "info": DesignToken,
        "infoDark": DesignToken,
        "warning": DesignToken,
        "warningDark": DesignToken,
        "success": DesignToken,
        "successDark": DesignToken
      }
    },
    "surface": {
      "overlay": DesignToken
    },
    "text": {
      "primary": DesignToken,
      "reversed": DesignToken
    },
    "header": {
      "global": {
        "primary": DesignToken,
        "reversed": DesignToken
      }
    }
  }
}